feat(types): migrate validator extras to TS
